LayoutTests/imported/w3c/web-platform-tests/inert/inert-computed-style-expected.txt


FAIL inert isn't hit-testable, but that isn't expose in the computed style assert_equals: inert node doesn't change user-select computed value expected (string) "auto" but got (undefined) undefined
